total cost + pages for orderItems and available products

+++ src/main/java/com/academy/course/appcafe/dto/OrderDTO.java
@@ -19,7 +19,7 @@ public class OrderDTO {
 
     private String paymentData;
 
-    private BigDecimal totalCost;
+    private Double totalCost;
 
     private Boolean isBought;
